Fix caching strategy for image builds
Change caching strategy for image builds. We used "registry" but we
switch to "inline" which is better because we do not need to
have separate "cache" tag for the image, caching metadata
is stored in the image manifests, which means that for the build
we can use the cache in the image itself when we build image locally.

This also fixes '--link' layers the images if cache is not available.

Previously we changed the ownership of the scripts and .local folder
to be airflow user but this was not really needed and even harmful as
those scripts could be modified by `airflow` user (even by mistake).

This caused problems when the scripts were used from the cache,
because when --link is used to copy the layers the user might
not be created yet. It turns out that you shoud never have
--link and --chown for the user that is created in previous steps.

The scripts are now owned by root users and the ".local" layer
that should be owned by the airflow user is now copied without
--link flag.